LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/hit-test/z-index-expected.txt


FAIL elementFromPoint should hit the last element in dom order. assert_equals: expected Element node <div id="b"></div> but got Element node <canvas id="canvas" width="200" height="200" layoutsubtre...
FAIL elementFromPoint should hit the highest z-index element. assert_equals: expected Element node <div id="a" style="z-index: 5;"></div> but got Element node <canvas id="canvas" width="200" height="200" layoutsubtre...
